Move clml.data.r-datasets-package to its own .ASD file
Quicklisp was not finding this system otherwise.

I am not sure why that is a problem. The system is referenced as a
dependency by data/r-datasets/clml.data.r-datasets.asd which seems to
have already defined this system in that same file. However I chose to
make this simple change rather than deep dive into Quicklisp and ASDF.